feat: Add Lua filter for RST notes (#88)
Fixes #76

Pandoc doesn't convert RST notes/warnings to their Confluence equivalents. I filed
https://github.com/jgm/pandoc/issues/10561 about it, but it seems difficult to fix. For
one thing, Confluence's `{note}` and `{info}` macros that we're using here, seem to be a
Confluence-specific thing. But the pandoc output we're using is called `jira`, and I
don't know if it's meant to be just Confluence or all Atlassian products. If the latter,
then we need more testing, with e.g. JIRA. These conversations and decisions seem bigger
than what we're fixing here.

And that's on top of the actual implementation details of the fix. Hopefully pandoc's
internal model supports these notes and warnings and we just need to map them to the RST
and Confluence equivalents.

But we can simplify our life here. There is a suggestion in that ticket to use a Lua
filter, and that's what this commit implements.

There is a new `rst_note_warning.lua` filter that turns RST notes (resp. warnings) into
Confluence `{info}` (resp. `{note}`). These mappings are not the most intuitive, but
they look more correct in Confluence.